data-center/passerelle-rf2/requirements.txt
requests
pyserial
paho-mqtt